requirements.txt
pelican==3.5.0
python-nvd3==0.13.5
pandas==0.15.2